Move Next unread button above the chat composer
